+++ src/rules/matches.js
@@ -1,8 +1,14 @@
+/**
+ * Matches Rule
+ * ------------
+ * Checks that the value of the field being validated matches the value of a separate specified field
+ *
+ */
 export default {
     condition: field => field.hasAttribute('data-val-equalto'),
 
     test: field => {
-        const matchedFieldName = field.getAttribute('data-val-equalto').replace('*.', '');
+        const matchedFieldName = field.getAttribute('data-val-equalto');
 
         if (matchedFieldName) {
             const input = document.querySelector(`input[name=${matchedFieldName}]`);